Add test for stub in DataService
Added the configurations necessary for the unit tests to function with
Cocoapods and then made a simple test for the stub method implemented in
the previous commit, which will soon request from the server across the
network to determine if server communication is working without having
to launch the application. NOTE that tests do not work with the actual
device, but only with the simulator, courtesy of Apple.