Error handle for requests
Before now, there wasn't correct error handling for the optionals
returned from the JSON objects that were being given from the server.
If the server is not on, this caused the app to crash. That is now
fixed, along with a quirk where arrays are call by value in swift, and
thus couldn't be edited in an outside context inside a callback. For
now, I'm using a container class (which uses call by reference) to
handle this until I find a more elegant solution.